Initial commit, including Makefile and execute-process
execute-process is a module that provides several overloaded functions for
running a child process.  The process is specified by a string and can
be passed arguments, can have input our output redirected, and can have
a time limit imposed on it.  The function determines how much time the
process has spent, much like the time command, and returns this information
as well as the process's exit code.