Create evaluate.cpp, the file containing main()
This file is responsible for evaluating command-line arguments, and calling
appropriate functions to evaluate the program.  As of this first commit,
evaluation means nothing more than running the program.  Eventually,
however, it should test for program correctness and speed based on external
files specified on the command-line.  Or, better yet, it should be able to
test on many files when a directory, and perhaps some extra information,
is given.