Add headers for classes Timer and Tester
These classes will handle reporting of the two halves of this program.
Timer will have different functions based on how detailed the reports are
to be.  It can output one test, a set of runs of the same test, or a set of
multiple tests.  It can report all of them, and/or the average.
Tester compares two files.  It can be run normally, which simply returns
true or false based on whether the files are the same.  Or, it can run
verbosely, which returns a string detailing the differences (or the empty
string if there are none).  It will also be able to be set to ignore
certain small differences.