Add documentation to Timer module
The new comments better describe the module.  In the .h file, they describe
how to use the class.  There is also a to-do list of features to add later.
In the .cpp file, they describe the libraries the implementation depends
upon, as well as a to-do list for implementation changes to make.