Add documentation to Tester module
The new comments better describe the module.  In the .h file, they describe
how to use the class, including available options and methods of running,
as well as describing the roles of the comparison and benchmark files.
In the .cpp file, it describes dependencies as well as an overview of the
algorithm.
Both files also have a to-do list of features to add in the future.