Update Makefile for new program name and organization
The program is no longer in test mode, and does far more than testing.
Therefore, it has been renamed  evaluate  to better match its function, as
used in the cpp file of its main().
It is also now better parameterized, with all needed files listed at top
in a variable, and all needed libraries similarly in a variable.