Add --spacing program option for styling timing output
This option changes the number of spaces used between columns.
A new method and member were added to the Timer class to accomplish this.
This member is used when outputting to determine how many spaces to output.
The repeat_char private method now no longer prints the character; it instead
returns a string, which allows its return value to, eg., be passed to a
function.