Modify header printed on timing tests
The Timer class now looks at the following of the current test, printing the
first that is nonempty:
1) Test name
2) Input file
3) Output file
4) Error file
5) "Unknown test"
evaluate.cpp now always sets the test name, setting it equal to the filename
(not the full extension) of the input file.
In the future, users may be allowed to specify test names to print, however
this seems like it would be difficult when the tests are not specified
one-by-one by the user.