Send recieved signals to child process
Added simple signal handling to the execute-process module, which, if possible,
sends the recieved signal to the child.
Currently, this handler is only set for SIGTERM and SIGINT, as these are the
ones considered likely for the user to use.  Now, for example, the user can
terminate an infinite loop in their program with ctrl-C, without having to
terminate the testing program as well.
As with timeouts, the result of a terminated program is neither tested
nor accounted for in the timing report.