Enhance determination of spacing in Timer class
Previously, it was assumed that fewer than 10 runs would be made for each test,
but that ought not to be assumed.  When more runs are made, the heading for
each column would be longer than expected by one or more digits.
To fix this, the num_digits() function is used, and that amount is subtracted
from the number of spaces to print.  The columns should thus line up for an
arbitrary number of tests.