Enhance printing during testing
Changed the printing of tests in --loud mode.  Previously, the test name was
printed twice.  Now it is only printed once, and is printed before the test
is run in every non-quiet mode.
Other print statements have been rearranged to better match with this.
The Tester class's print_on_success option is no longer utilized.  Instead,
the evaluate module checks manually whether to print or not based on the
options set.
Finally, added a statement at the end enumerating the number of tests passed
out of the total number of tests.  This prints in all modes.